updating scripts

+++ Shell/osc.sh
@@ -3,6 +3,9 @@ function atomup {
 	unset verl
 	cd $SPEC
 	if [[ $USER == makerpm ]]; then
+		if ! [[ -d $HOME/atom ]]; then
+			git clone https://github.com/atom/atom $HOME/atom
+		fi
 		pushd ~/atom
 		git checkout stable
 		git pull origin stable
@@ -24,26 +27,28 @@ function oscb {
 }
 
 function vimup {
-	unset verc
-	unset verl
-	unset VIMDIR
-	pushd ~/GitHub/vim # change into my local Vim git repo copy
-	git pull origin master # Update local repo
-	# verc is the current patch version of Vim in vim.spec
-	# verl is the latest patch version of Vim.
-	VIMDIR=/home/fusion809/OBS/home:fusion809/vim
-	verc=$(sed -n 's/%define patchlevel\s\s*//p' $VIMDIR/vim.spec)
-	verl=$(git describe --tags | sed 's/^v7.4.//;s/-/./g')
-	popd
-	pushd $VIMDIR
-	if [[ $verc == $verl ]]; then
-		echo "Vim is up-to-date"
-	else
-		sed -i -e "s/%define patchlevel  $verc/%define patchlevel  $verl/g" vim.spec
-		osc rm *.gz
-		wget -c https://github.com/vim/vim/archive/v7.4.$verl.tar.gz
-		osc add v7.4.$verl.tar.gz
-		osc ci -m "Updating to version 7.4.$verl"
+	if [[ $USER == fusion809 ]]; then
+		unset verc
+		unset verl
+		unset VIMDIR
+		pushd ~/GitHub/vim # change into my local Vim git repo copy
+		git pull origin master # Update local repo
+		# verc is the current patch version of Vim in vim.spec
+		# verl is the latest patch version of Vim.
+		VIMDIR=/home/fusion809/OBS/home:fusion809/vim
+		verc=$(sed -n 's/%define patchlevel\s\s*//p' $VIMDIR/vim.spec)
+		verl=$(git describe --tags | sed 's/^v7.4.//;s/-/./g')
+		popd
+		pushd $VIMDIR
+		if [[ $verc == $verl ]]; then
+			echo "Vim is up-to-date"
+		else
+			sed -i -e "s/%define patchlevel  $verc/%define patchlevel  $verl/g" vim.spec
+			osc rm *.gz
+			wget -c https://github.com/vim/vim/archive/v7.4.$verl.tar.gz
+			osc add v7.4.$verl.tar.gz
+			osc ci -m "Updating to version 7.4.$verl"
+		fi
+		popd
 	fi
-	popd
 }
